moving client related code to its own directory. Splitting off encoding functionality from the paginator (hopefully to be configurable one day
